requirements.txt
fastapi
pymongo
python-dotenv
uvicorn
motor
beanie
passlib[bcrypt]
python-multipart
python-jose[cryptography]
cloudinary